Restructure commit to isolate Chrome extension
This new layout allows various outputs to be created.  In the future
more browsers will have extension APIs and we need to think about
sharing as much code as possible.

I've also removed the old Chrome extension plugin for Grunt as it does
not allow for scattered files to unify into an extension.