Minor cleanup, file removal and test fixes
The FireFox tests were not running after Chrome, but still reporting a
pass.  It seems that Grunt is conflating test files and not running them
a second time.  This problem appears with both simple-mocha and
mocha-test tasks.  I've worked around the problem for now by running
each grunt task in isolation, this is less than optimal so further
research will be required to resolve the issue.